updated spell detail and order detail to pull directly from id routes in API, link to spell detail page
