+++ client/store/index.js
@@ -3,7 +3,8 @@ import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
-import spells, {currentSpellReducer} from './spell'
+import spells from './spell'
+import currentSpell from './spell-detail'
 import cart from './cart'
 import orders from './order'
 import userOrders from './user-orders'
@@ -16,7 +17,7 @@ const reducer = combineReducers({
   orders,
   userOrders,
   orderSpells,
-  currentSpellReducer,
+  currentSpell,
 })
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
@@ -26,6 +27,7 @@ const store = createStore(reducer, middleware)
 export default store
 export * from './user'
 export * from './spell'
+export * from './spell-detail'
 export * from './cart'
 export * from './order'
 export * from './user-orders'
